backend/requirements.txt
django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.3.0
django-cors-headers==4.0.0
PyJWT==2.8.0
asgiref ==3.8.1
setuptools==59.6.0
sqlparse==0.5.0
typing_extensions==4.12.0
django-cors-headers==4.4.0
numpy==2.1.0


